requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# IR libraries
rank-bm25>=0.2.2
gensim>=4.3.0

# NLP
nltk>=3.8.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
